# openapi/schema.json
{
  "openapi": "3.0.0",
  "paths": {
    "/courses": {
      "get": {
        "operationId": "CoursesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchText",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfCourseDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CourseDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Find all courses",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/physiotherapist": {
      "get": {
        "operationId": "CoursesController_findAllForPhysiotherapist",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchText",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfCourseDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CourseDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/subscribed/{userId}": {
      "get": {
        "operationId": "CoursesController_findSubscribedCoursesForUserId",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchText",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfCourseDto",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CourseDto"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find all courses to which the given user is subscribed",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "summary": "Find a course by its id",
        "tags": [
          "Courses"
        ]
      },
      "put": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a course by its id",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a course by its id",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/picture": {
      "post": {
        "operationId": "CoursesController_uploadCoursePicture",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a course picture",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/subscriptions/course/{courseId}": {
      "get": {
        "operationId": "SubscriptionsController_getSubscribersForCourse",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfSubscriptionDtoWithUser",
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SubscriptionDtoWithUser"
                          }
                        }
                      },
                      "required": [
                        "data"
                      ]
                    },
                    {
                      "properties": {
                        "meta": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "number"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "currentPage": {
                              "type": "number"
                            },
                            "totalItems": {
                              "type": "number"
                            },
                            "totalPages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "items",
                            "hasNextPage",
                            "currentPage",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "meta"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN & PHYSIOTHERAPIST] Get all active subscribers for a course",
        "tags": [
          "Subscriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "EasyMotion API",
    "description": "The EasyMotion API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "Enter your Bearer token"
      }
    },
    "schemas": {
      "CourseOwnerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID)",
            "example": "b3bf4d18-8dd0-43a1-b1da-fd3f7b9553a1"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName"
        ]
      },
...
      "CourseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the course"
          },
          "image_path": {
            "type": "string",
            "nullable": true,
            "description": "The image path of the course"
          },
          "name": {
            "type": "string",
            "description": "The name of the course"
          },
          "description": {
            "type": "string",
            "description": "The complete description of the course"
          },
          "short_description": {
            "type": "string",
            "description": "A brief description of the course"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Location where the course will be held"
          },
          "instructors": {
            "description": "List of instructors for the course",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "level": {
            "type": "string",
            "description": "Level of the course",
            "enum": [
              "BASIC",
              "MEDIUM",
              "ADVANCED"
            ]
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "Price of the course"
          },
          "payment_recurrence": {
            "type": "string",
            "description": "Payment recurrence",
            "enum": [
              "SINGLE",
              "PER_SESSION",
              "MONTHLY",
              "ANNUAL"
            ]
          },
          "is_published": {
            "type": "boolean",
            "description": "Indicates if the course is public"
          },
          "subscriptions_open": {
            "type": "boolean",
            "description": "Indicates if subscriptions are open"
          },
          "max_subscribers": {
            "type": "number",
            "nullable": true,
            "description": "Amount of maximum subscribers (unlimited if null)"
          },
          "current_subscribers": {
            "type": "number",
            "description": "Current subscribers (count)"
          },
          "tags": {
            "description": "Related tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date of creation"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date of last update"
          },
          "owner": {
            "description": "The owner of the course",
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseOwnerDto"
              }
            ]
          },
          "category": {
            "description": "The category of the course",
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseCategoryDto"
              }
            ]
          },
          "sessions": {
            "description": "The sessions of the course",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseSessionDto"
            }
          },
          "subscription_start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Subscription start date"
          },
          "subscription_end_date": {
            "format": "date-time",
            "type": "string",
            "description": "Subscription end date"
          }
        },
        "required": [
          "id",
          "image_path",
          "name",
          "description",
          "short_description",
          "instructors",
          "level",
          "price",
          "payment_recurrence",
          "is_published",
          "subscriptions_open",
          "current_subscribers",
          "tags",
          "created_at",
          "updated_at",
          "owner",
          "category",
          "sessions",
          "subscription_start_date",
          "subscription_end_date"
        ]
      },
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the course"
          },
          "description": {
            "type": "string",
            "description": "The complete description of the course"
          },
          "short_description": {
            "type": "string",
            "description": "A brief description of the course"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Location where the course will be held"
          },
          "instructors": {
            "description": "List of instructors for the course",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "level": {
            "type": "string",
            "description": "Level of the course",
            "enum": [
              "BASIC",
              "MEDIUM",
              "ADVANCED"
            ]
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "Price of the course"
          },
          "payment_recurrence": {
            "type": "string",
            "description": "Payment recurrence",
            "enum": [
              "SINGLE",
              "PER_SESSION",
              "MONTHLY",
              "ANNUAL"
            ]
          },
          "is_published": {
            "type": "boolean",
            "description": "Indicates if the course is public"
          },
          "subscriptions_open": {
            "type": "boolean",
            "description": "Indicates if subscriptions are open"
          },
          "max_subscribers": {
            "type": "number",
            "nullable": true,
            "description": "Amount of maximum subscribers (unlimited if null)"
          },
          "tags": {
            "description": "Related tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscription_start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Subscription start date"
          },
          "subscription_end_date": {
            "format": "date-time",
            "type": "string",
            "description": "Subscription end date"
          },
          "category_id": {
            "type": "string",
            "description": "id of the existent category"
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCourseSessionDto"
            }
          }
        },
        "required": [
          "name",
          "description",
          "short_description",
          "instructors",
          "level",
          "price",
          "payment_recurrence",
          "is_published",
          "subscriptions_open",
          "tags",
          "subscription_start_date",
          "subscription_end_date",
          "category_id",
          "sessions"
        ]
      },
...
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the course"
          },
          "description": {
            "type": "string",
            "description": "The complete description of the course"
          },
          "short_description": {
            "type": "string",
            "description": "A brief description of the course"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Location where the course will be held"
          },
          "instructors": {
            "description": "List of instructors for the course",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "level": {
            "type": "string",
            "description": "Level of the course",
            "enum": [
              "BASIC",
              "MEDIUM",
              "ADVANCED"
            ]
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "Price of the course"
          },
          "payment_recurrence": {
            "type": "string",
            "description": "Payment recurrence",
            "enum": [
              "SINGLE",
              "PER_SESSION",
              "MONTHLY",
              "ANNUAL"
            ]
          },
          "is_published": {
            "type": "boolean",
            "description": "Indicates if the course is public"
          },
          "subscriptions_open": {
            "type": "boolean",
            "description": "Indicates if subscriptions are open"
          },
          "max_subscribers": {
            "type": "number",
            "nullable": true,
            "description": "Amount of maximum subscribers (unlimited if null)"
          },
          "tags": {
            "description": "Related tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscription_start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Subscription start date"
          },
          "subscription_end_date": {
            "format": "date-time",
            "type": "string",
            "description": "Subscription end date"
          },
          "category_id": {
            "type": "string",
            "description": "id of the existent category"
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateCourseSessionDto"
            }
          }
        }
      },
...
      "SubscriptionDtoWithUser": {
        "type": "object",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Update date"
          },
          "subscriptionRequestMessage": {
            "type": "string",
            "nullable": true,
            "description": "Subscription request message"
          },
          "course": {
            "description": "Course id and name",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionCourseDto"
              }
            ]
          },
          "user": {
            "description": "Subscribed user",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionUserDto"
              }
            ]
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "course",
          "user"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}